docs(api): 更新马来供应商文档，添加通用响应定义及示例

+++ src/main/java/com/heytrip/hotel/supplier/entity/ApiCallLog.java
@@ -137,7 +137,7 @@ public class ApiCallLog {
      */
     @Column(name = "retry_count")
     @Comment("重试次数")
-    private Integer retryCount = 0;
+    private Long retryCount = 0L;
     
     /**
      * 是否成功：1-成功，0-失败
